refactor: move some functions from ResetPasswordService into User model
