+++ app/models/user.ts
@@ -3,6 +3,7 @@ import { DateTime } from "luxon";
 import { withAuthFinder } from "@adonisjs/auth/mixins/lucid";
 import { compose } from "@adonisjs/core/helpers";
 import hash from "@adonisjs/core/services/hash";
+import logger from "@adonisjs/core/services/logger";
 import { BaseModel, beforeSave, column, scope } from "@adonisjs/lucid/orm";
 
 import { sha256 } from "#utils/hash";
@@ -57,5 +58,43 @@ export default class User extends compose(BaseModel, AuthFinder) {
       this.resetPasswordTokenExpiration !== null &&
       this.resetPasswordTokenExpiration >= DateTime.now()
     );
+  }
+
+  /**
+   * Updates the user's password, invalidates any existing reset tokens and saves the model
+   *
+   * @param password new password
+   */
+  async resetPassword(password: string) {
+    this.password = password;
+    await this.destroyToken();
+    logger.info(`Password changed for user ${this.email}`);
+  }
+
+  /**
+   * Invalidates any existing reset tokens and saves the model
+   */
+  async destroyToken() {
+    this.resetPasswordToken = null;
+    this.resetPasswordTokenExpiration = null;
+    await this.save();
+  }
+
+  /**
+   * Generates random secure token for password reseting and stores it in database
+   * Hashing before storing should be handled by `beforeSave()` hook in lucid
+   * Note: saves the model
+   *
+   * @param expirationDate time till the token is valid. Pass DateTime instance using e.g. `DateTime.now().plus({minutes: 15})`
+   * @returns generated token (not hashed)
+   */
+  async generateResetPasswordToken(
+    expirationDate: DateTime<true>,
+  ): Promise<string> {
+    const token = crypto.randomUUID().toString();
+    this.resetPasswordToken = token;
+    this.resetPasswordTokenExpiration = expirationDate;
+    await this.save();
+    return token;
   }
 }
